+++ api/src/backend/api/v1/serializers.py
@@ -1525,6 +1525,11 @@ class BaseWriteProviderSecretSerializer(BaseWriteSerializer):
                             "or both 'api_key' and 'api_email'."
                         }
                     )
+            elif provider_type == Provider.ProviderChoices.OPENSTACK.value:
+                if "clouds_yaml_content" in secret:
+                    serializer = OpenStackCloudsYamlProviderSecret(data=secret)
+                else:
+                    serializer = OpenStackExplicitProviderSecret(data=secret)
             else:
                 raise serializers.ValidationError(
                     {"provider": f"Provider type not supported {provider_type}"}
@@ -1686,6 +1691,27 @@ class CloudflareTokenProviderSecret(serializers.Serializer):
 class CloudflareApiKeyProviderSecret(serializers.Serializer):
     api_key = serializers.CharField()
     api_email = serializers.EmailField()
+
+    class Meta:
+        resource_name = "provider-secrets"
+
+
+class OpenStackCloudsYamlProviderSecret(serializers.Serializer):
+    clouds_yaml_content = serializers.CharField()
+    clouds_yaml_cloud = serializers.CharField()
+
+    class Meta:
+        resource_name = "provider-secrets"
+
+
+class OpenStackExplicitProviderSecret(serializers.Serializer):
+    auth_url = serializers.CharField()
+    username = serializers.CharField()
+    password = serializers.CharField()
+    region_name = serializers.CharField()
+    identity_api_version = serializers.CharField(required=False)
+    user_domain_name = serializers.CharField(required=False)
+    project_domain_name = serializers.CharField(required=False)
 
     class Meta:
         resource_name = "provider-secrets"